test(RateLimiter): consolidate resetAfter coverage
